refactored: Customer and User are supposed to be same thing
Customer and User should be the same thing.  I merged the 2 models into one.
