+++ backend/src/main/java/com/lambdaschool/coffeebean/repository/Customerrepository.java
@@ -1,8 +1,8 @@
-package com.lambdaschool.coffeebean.repository;
-
-import com.lambdaschool.coffeebean.model.Customer;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface Customerrepository extends JpaRepository<Customer, Long>
-{
-}
+//package com.lambdaschool.coffeebean.repository;
+//
+//import com.lambdaschool.coffeebean.model.Customer;
+//import org.springframework.data.jpa.repository.JpaRepository;
+//
+//public interface Customerrepository extends JpaRepository<Customer, Long>
+//{
+//}
